+++ ipld/unixfs/io/directory_test.go
@@ -990,3 +990,4 @@ func TestHAMTDirectorySizeEstimationMode(t *testing.T) {
 		assert.Equal(t, SizeEstimationBlock, hamtDir.GetSizeEstimationMode())
 	})
 }
+